LayoutTests/imported/w3c/IndexedDB-private-browsing/idbfactory_deleteDatabase2-expected.txt


PASS IDBFactory.deleteDatabase() - result of the request is set to undefined
